refactor: refactor action component
